+++ apps/openmw/mwmechanics/alchemy.hpp
@@ -86,7 +86,7 @@ namespace MWMechanics
         int countPotionsToBrew() const;
         ///< calculates maximum amount of potions, which you can make from selected ingredients
 
-        static bool knownEffect(unsigned int potionEffectIndex, const MWWorld::Ptr& npc);
+        static bool knownEffect(size_t potionEffectIndex, const MWWorld::Ptr& npc);
         ///< Does npc have sufficient alchemy skill to know about this potion effect?
 
         void setAlchemist(const MWWorld::Ptr& npc);
